refactor: 투표 상세조회 리팩토링
