+++ api/src/main/java/com/programmers/bucketback/domains/vote/application/VoteService.java
@@ -6,6 +6,7 @@ import com.programmers.bucketback.Hobby;
 import com.programmers.bucketback.common.cursor.CursorPageParameters;
 import com.programmers.bucketback.common.cursor.CursorSummary;
 import com.programmers.bucketback.domains.vote.application.dto.request.VoteCreateServiceRequest;
+import com.programmers.bucketback.domains.vote.application.dto.response.VoteGetServiceResponse;
 import com.programmers.bucketback.domains.vote.domain.Vote;
 import com.programmers.bucketback.domains.vote.implementation.VoteAppender;
 import com.programmers.bucketback.domains.vote.implementation.VoteManager;
@@ -14,7 +15,6 @@ import com.programmers.bucketback.domains.vote.implementation.VoteRemover;
 import com.programmers.bucketback.domains.vote.model.VoteCursorSummary;
 import com.programmers.bucketback.domains.vote.model.request.VoteSortCondition;
 import com.programmers.bucketback.domains.vote.model.request.VoteStatusCondition;
-import com.programmers.bucketback.domains.vote.model.response.VoteGetServiceResponse;
 import com.programmers.bucketback.error.exception.BusinessException;
 import com.programmers.bucketback.error.exception.ErrorCode;
 import com.programmers.bucketback.global.util.MemberUtils;
@@ -74,8 +74,9 @@ public class VoteService {
 
 	public VoteGetServiceResponse getVote(final Long voteId) {
 		Long memberId = MemberUtils.getCurrentMemberId();
+		VoteCursorSummary summary = voteReader.read(voteId, memberId);
 
-		return voteReader.read(voteId, memberId);
+		return VoteGetServiceResponse.from(summary);
 	}
 
 	public CursorSummary<VoteCursorSummary> getVotesByCursor(
